agents-adk-docs/adk-plugins/requirements.txt
google-adk
prometheus_client
